refactor(model): improve Tickets class modularity and readability - Extract ticket generation into TicketGenerator class - Extract SECOND rank check into isSecondRank method
